product-service/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
python-dotenv
elasticsearch==8.11.1
pika